requirements.txt
appscript==1.1.0
argh==0.26.2
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coloredlogs==10.0
gitdb2==2.0.6
GitPython==3.0.5
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
httplib2==0.15.0
humanfriendly==4.18
idna==2.8
oauth2client==4.1.3
pathtools==0.1.2
Pillow==6.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
rsa==4.0
six==1.13.0
smmap2==2.0.5
uritemplate==3.0.1
urllib3==1.25.7
watchdog==0.9.0
